Breaking change: replace --no-tlog-upload flag with --tlog-upload flag
In 1.0, users could upload to the tlog and use a signing key by setting the COSIGN_EXPERIMENTAL env var.

Now that the env var is being removed, we still need to give users a way to sign with a key and upload to the tlog.

Replacing --no-tlog-upload with --tlog-upload gives us this opportunity, and is more intuitive.

Signed-off-by: Priya Wadhwa <priya@chainguard.dev>