Breaking change: Respect tlog-upload=false, default to true (#2505)
* Respect tlog-upload=false, default to true

This changes the default for uploading to the tlog to always upload
unless explicitly set to false. We no longer require checking if the
flow is keyless or not, as it'll be up to the uploader to decide if they
want to upload to Rekor or not.

When verifying, if you set tlog-upload=false, then you must set
insecure-skip-tlog-verify=true

BREAKING CHANGE: Require tlog-upload=false to not upload to Rekor. By default, signatures will be uploaded.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>