Fix timestamp verification, add verify-blob tests (#2527)
* Fix timestamp verification, add verify-blob tests

While adding unit tests for verify-blob with timestamping, I found that
we try to verify with the current time if no bundle is provided, even if
we have a timestamp. I cleaned up this function to check the timestamp
time and bundle time if either are provided, and if neither is provided,
check against the current time.

I also added a mock client to clean up tests. I'll move this mock client
over to the timestamp-authority repo.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>

* Remove println

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>