Fix verifying when a certificate is provided
Keys were preferred over certificates. This caused an issue, because in
the CLI functions that call verify, if you pass a certificate to verify
with, we extract the public key from it and set that to co.SigVerifier
while also setting sig.Cert. Checking if a certificate was provided
first should mitigate this.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>