Add environment variable token provider.
This adds a new token provider that looks for tokens located at SIGSTORE_ID_TOKEN.

This is an alternative to COSIGN_IDENTITY_TOKEN, which is implemented as
a Viper binding.

ID_TOKEN was used instead of IDENTITY_TOKEN, since this seems to be
consistent with other providers (e.g. ACTIONS_ID_TOKEN_REQUEST_TOKEN).
If this ends up being confusing we can add support for both variants
later if needed.

Signed-off-by: Billy Lynch <billy@chainguard.dev>