ggcr: Use remote.Reuse to make everything faster
Previously, every call to a remote.Function was performing a Ping and
Token exchange, which meant that we were doing 3 round trips instead of
1 (for most calls).

The remote.Reuse option effectively tells the remote package to call
methods on a Pusher/Puller instance rather than re-doing the Ping and
Token requests.

The alternative to this PR would be to plumb around a puller and pusher
instance and refactor all of cosign to call methods on those, but this
is much simpler.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>