chore(refactor) refactors error handling with exit codes
previously we had a lookup table with error types and error messages
that meant that to add an error and exit code, a user had to add the
type, the message and the exit code mapping as well as modify the
location where it was thrown. Although this hasn't largely changed, we
were having to force/shove specific error messages into the old
VerificationError struct. Now, we have removed the error types and
error messages and instead define types for each of the errors
allowing us to flexibly customise each error type if needed. This
allows us to make use of errors.Is for exit code lookups also.

Signed-off-by: ChrisJBurns <29541485+ChrisJBurns@users.noreply.github.com>