Allow overriding remote options (#3049)
RegistryOptions.GetRegistryClientOpts is used in a bunch of places, and
it always instantiates a new set of options. This is ~fine for callers
that only use the options once (e.g. the cosign CLI), but is unfortunate
for callers that would like to reuse those options.

It's useful to reuse these options because they represent a shared
context that eliminates a lot of duplicate work across calls
(e.g. authenticating with the registry and doing existence checks).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>