Fail timestamp verification if no root is provided
The bug was that we would conditionally check a timestamp if a root was
provided. If no root was provided even if a timestamp was provided, then
signature verification would succeed.

The good news is this will not show a successful signature if the
transparency log does not contain the entry too, for timestamp
verification.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>