Add support for OpenVEX predicate type
OpenVEX is an implementation of the Vulnerability Exploitability
Exchange (VEX) designed to be attestable, SBOM-agnostic and lightweight.
It is hosted in the OpenSSF Vulnerability Disclosures WG and has
support in popular scanners such as Trivy adn Grype.

This PR adds support for openvex predicates to the `cosign attest` and
`cosign download attestation` commands.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>