Use local rekor and fulcio instances in e2e tests (#3478)
* Clean up e2e test script

- Use -v for clearer output
- All the tests tagged 'e2e' are in ./test, no need to search the whole
  tree
- There's no third_party directory, no need to grep it out
- Clean up services no matter when the script exits

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* test: Clean up cosign-secret

Clean up the key pair secret after the test so that the test suite can
be re-run locally if desired. Without this, the secret is set to
'immutable' and can't be overwritten by the next test run.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Use local rekor and fulcio instances in e2e tests

In 70683571 the e2e tests moved from running on the locally-spun-up
rekor instance to the public instance. This means test signatures are
piling up in the public instance, and the tests may be taking longer
than they need to since they are using an external service.

This change moves back to using the local rekor instance, which the e2e
has still been spinning up even though it has been going unused. Also
now do the same for fulcio.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

---------

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>