Don't ignore transparency log in tests if possible
Update the e2e tests to default to setting IgnoreTlog to false where
possible. In some cases, where the IgnoreTlog functionality is being
explicitly tested, continue to set it to true.

Since the transparency log isn't being ignored, the signing commands
need to upload it and need the rekor public key and URL in order to do
so.

Removes one redundant test.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>