Remove cross.yaml workflow (#3629)
The artifacts uploaded by cross.yaml are not used anywhere. Moreover,
the ability to build on all three platforms and use the resulting binary
is already tested in e2e-with-binary.yml. This change removes the
workflow for the sake of decluttering the workflows and reducing our use
of GitHub storage.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>