Add debug providers command.
When trying to run cosign on a remote workstation it would be nice to
have a command to confirm the OIDC provider behavior of what is/isn't
enabled and in what order.

This PR adds a new hidden debug subcommand, with a debug providers
subcommand that prints out the list of orderd providers and whether
or not they are enabled.

Signed-off-by: Billy Lynch <billy@chainguard.dev>