Make config layers in ociremote mountable
The wrapping that oci.Signatures does hides the ConfigLayer()
implementation in remote.Image that remote.Write relies on for
determining if it can mount the config blob (vs having to re-upload it),
so remote.Write is sending additional blob uploads, which incurs an
additional roundtrip that we don't really need to do.

Explicitly implement ConfigLayer() in the wrappers by dispatching to the
wrapped implementation fixes this.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>