Fix for multiple WithRemote options (#3982)
When specifying multiple sets of remote options using
WithRemoteOptions, only the last set of options are used.
The earlier ones are discarded. This is a problem for example, when
an initial set of WithRemoteOptions are generated by command-line
flags, and then additional ones need to be specified during
application execution. Doing so would discard the options set
by the command-line flags.

To preserve backwards compatibility, where the WithRemoteOptions
function overrides the defaults, an additional function
WithMoreRemoteOptions has been added, which appends the options
to any existing ones.

Resolves: sigstore/cosign/#3981

Signed-off-by: Warren Hodgkinson <warren.hodgkinson@gmail.com>