Update SHA-1 CT certificates in tests
SHA-1 in crypto/x509 has been removed as of Go 1.24, so our tests are
failing. This updates the sct verification tests to use certificates
that are hashed with SHA-256 or above.

Unfortuantely, regenerating inclusion proofs independent of certificates
is quite tricky, so I've chosen to drop tests for detached SCTs. This
is fine because the public-good Fulcio does not produce detached SCTs,
no other Sigstore clients support detached SCTs, and we've announced
that Fulcio will remove this feature in the next revision.

Fixes https://github.com/sigstore/cosign/issues/2091

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>