Make e2e service health check more flexible
In 62a6617 in rekor, health checks were added to the trillian services,
which means now five of the five services will report "healthy" when
running. The cosign e2e tests were checking that exactly 3 were healthy,
so if they started too fast and 4 or more were healthy, the tests would
fail. This change makes the check more flexible. It is not very
sophisticated, as it's technically possible the wrong 3 services are
healthy, but it should still be good enough to serve its purpose in CI.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>