Fix cert verification logic for trusted-root/SCTs (#4294)
Previously, the logic for verifying the Fulcio leaf cert was gated on
whether the verification material contained a trusted root and whether
the cert had an embedded SCT. This was because the upgraded TUF client
that fetched the trusted root and passed it to the sigstore-go
verification routines was developed concurrently with updates to
sigstore-go[1], and sigstore-go didn't initially support returning the
cert chain that would allow verifying a detached SCT. When the change in
sigstore-go landed, the API call was updated but the logic and the
comment was not fixed. This meant that a trusted root file could not be
used to verify leaf certs, unnecessarily. Since the trusted root was
set, the TUF client didn't fetch the Fulcio certs independently, so the
cert failed to verify. This change removes the restriction and adds a
test that uses trusted root.

[1] https://github.com/sigstore/sigstore-go/pull/300

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>