feat(cli): add `--payload-media-type` to `sign`
This commit introduces a new `--payload-media-type` flag to the
`cosign sign` command.

When signing with a custom payload using the `--payload` flag, users can
now also specify the MIME type of that payload via `--payload-media-type`.

This media type is then stored in the `mediaType` field of the signature
layer descriptor in the resulting OCI manifest. This allows consumers of
the signature to correctly interpret the payload's content type.

If `--payload` is not used, this flag is ignored.

Signed-off-by: Alberto Gonzalez <albgonz@google.com>