Verify subject with bundle only when checking claims (#4320)
Previously, when --check-claims was set to false and a bundle in the new
format was provided, we'd still try to check the in-toto subject digest
and algorithm. These values weren't being set since they were
conditioned on checking claims. Now, we skip digest verification if
check-claims is false with a new bundle.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>