Relax go directive to 1.24.x
One of the dependencies has hardcoded a specific version of go, which
forces all clients to use that version of Go in their own modules. This
is unnecessarily restrictive, rather libraries should specify the
minimum Go version necessary to build, and consumers should use the
latest patch release when building to pick up bug fixes.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>