Move loading key to common
Move the setting of SigVerifier based on the key ref, key slot, or cert
and cert chain, to the common file.

For verifying blobs and blob attestations with a certificate instead of
a key, we return the cert which is used directly in the options list for
verification. For images, the cert and cert chain must be validated and
then unpacked into the SigVerifier, where the cosign Verify* functions
check its validity by extracting it from the verifier.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>