Deprecate offline flag
The offline flag is misleading and is a no-op with the new Cosign v3
defaults. The flag's purpose was to prevent a client from falling back
to verifying an artifact's inclusion in Rekor when a proof failed to
verify. Most users thought offline verification forced the client to not
make any network requests - a very reasonable assumption, but with TUF,
network requests are a part of verification if the local TUF metadata
has expired.

I've updated the README as well, though we need to make a far more
comprehensive pass over the documentation since it's out of date given
our new trusted-root/bundle flags.

Fixes #4454

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>