Deduplicate code in sign/attest* and verify* commands (#4449)
* Deduplicate key/token handling in sign commands

Move the nearly identical code for parsing key options and creating a
key pair and token out of attest, attest-blob, sign, and sign-blob, and
into a common helper package. Move functions that had been shared out of
sign.go into the helper package too so that other commands do not have
to import the sign command package.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Deduplicate signer-verifier creation

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Deduplicate timestamp retrieval

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Deduplicate rekor upload

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Deduplicate bundle compilation

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Move OCI parsing function to signcommon

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Make flag compatibility checking consistent

Move flag checks when --new-bundle-format is used to a common helper
module and have all four verify commands use it. Remove redundant flag
checker code.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Remove duplicate certs setting

RootCerts and IntermediateCerts are already set on CheckOpts during
loadCertsKeylessVerification.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Move loading key to common

Move the setting of SigVerifier based on the key ref, key slot, or cert
and cert chain, to the common file.

For verifying blobs and blob attestations with a certificate instead of
a key, we return the cert which is used directly in the options list for
verification. For images, the cert and cert chain must be validated and
then unpacked into the SigVerifier, where the cosign Verify* functions
check its validity by extracting it from the verifier.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Deduplicate TUF v1 fetch and rekor client setup

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Deduplicate trusted material setting

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Move common functions to common.go

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

---------

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>